Fixed responsiveness and directory structure
