+++ client/src/app/auth/login/page.tsx
@@ -1,5 +1,5 @@
-"use client";
-import React, { useState } from "react";
+"use client"
+import { useState } from "react";
 import Link from "next/link";
 import { GoogleIcon } from "@/icons/GoogleIcon";
