+++ client/src/app/auth/signup/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { useState } from "react";
+import { useState } from "react";
 import Link from "next/link";
 
 type FormValues = {
