+++ client/src/app/globals.css
@@ -1,11 +1,6 @@
-@import url('https://fonts.googleapis.com/css2?family=Lexend:wght@100..900&display=swap');
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-* {
-  font-family: "Lexend";
-}
 
 
 :root {
