+++ client/src/app/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from "next";
 import { Poppins } from "next/font/google";
 import "./globals.css";
-import Navbar from "@/components/Nav/Nav";
-import Footer from "@/components/footer";
+import Navbar from "@/Components/Navbar";
+import Footer from "@/Components/Footer";
+import { UrlLogo } from "@/icons/UrlLogo";
 
 const poppins = Poppins({ subsets: ["latin"], weight: "400" });
 
@@ -10,7 +11,7 @@ export const metadata: Metadata = {
   title: "ShortURL",
   description: "Short URLs, Custom Links and Analytics",
   icons: {
-    icon: "/UrlLogo.svg",
+    
   },
 };
