--- client/src/app/page.tsx
@@ -1,11 +1,0 @@
-// pages/index.js or pages/home.js
-"use client";
-import HomePage from '@/components/home/page'
-
-
-export default function Home() {
-
-  return (
-    <HomePage />
-  );
-}
