--- client/src/components/About/about.tsx
@@ -1,19 +1,0 @@
-"use client"
-import Abone from './abone';
-import Abtwo from './abtwo';
-import Abthree from './abthree';
-import Abfour from './abfour';
-import Abfive from './abfive';
-
-const about = () => {
-    return (
-        <div>
-            <Abone/>
-            <Abtwo/>
-            <Abthree/>
-            <Abfour/>
-            <Abfive/>
-        </div>
-    )
-}
-export default about;
