--- client/src/components/Nav/Nav.css
@@ -1,25 +1,0 @@
-.hero {
-    padding: 15px;
-    border-radius: 20px;
-}
-
-.navOption {
-    position: relative;
-    display: inline-block;
-  }
-  .navOption::after {
-    content: '';
-    position: absolute;
-    width: 100%;
-    height: 2px;
-    bottom: 0;
-    left: 0;
-    background-color: currentColor;
-    transform: scaleX(0);
-    transform-origin: bottom right;
-    transition: transform 0.3s ease-out;
-  }
-  .navOption:hover::after {
-    transform: scaleX(1);
-    transform-origin: bottom left;
-  }
