--- client/src/components/home/home.css
@@ -1,18 +1,0 @@
-/* .custom-list {
-    list-style-type: none;
-    padding: 0;
-  }
-  
-  .custom-list li {
-    position: relative;
-    padding-left: 24px; 
-    margin-bottom: 8px; 
-  }
-  
-  .custom-list li::before {
-    content: '🔒';
-    position: absolute;
-    left: 0;
-    top: 0;
-  }
-   */
