--- client/src/components/home/page.tsx
@@ -1,14 +1,0 @@
-"use client"
-import One from './top';
-import Two from './mid';
-import Three from './bottom';
-
-export default function Home() {
-    return (
-        <div>
-            <One />
-            <Two />
-            <Three />
-        </div>
-    );
-}
