--- client/src/components/home/top.css
@@ -1,13 +1,0 @@
-
-  @media (min-width: 768px) {
-    #centered-placeholder::placeholder {
-      text-align: left;
-    }
-  }
-  
-  @media (max-width: 767px) {
-    #centered-placeholder::placeholder {
-      text-align: center;
-    }
-  }
-
